finish logging in before getting menu data
